511API/CrashDataService/requirements.txt
requests
configparser
geopandas
pandas
shapely
fiona
azure-functions
